refactor: restructured test_cases
